requirements.txt
psycopg2-binary==2.9.9
pandas==2.2.2
pyproj==3.6.1
openai==1.35.3
apscheduler==3.10.4
fastapi==0.111.0
uvicorn[standard]==0.30.1
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.9
cloudinary==1.40.0
httpx==0.27.0
fuzzywuzzy==0.18.0
python-Levenshtein==0.25.1
jaro-winkler==2.0.3
pgvector==0.3.2
pydantic==2.7.1
pydantic-settings==2.3.1
python-dotenv==1.0.1
aiohttp==3.9.5
tenacity==8.3.0
